requirements.txt
# Dependências do projeto
fastapi
uvicorn
pydantic
python-dotenv